pyproject.toml
[project]
name = "racing-adapter"
version = "0.1.0"
description = "HID device adapter for racing peripherals"
requires-python = ">=3.9"
dependencies = [
    "hidapi>=0.14.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "openai>=1.0.0",
    "pygame>=2.6.1",
    "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
    "pyserial>=3.5",
    "python-dotenv>=1.0.0",
]
